+++ src/entities/comment/libs/hooks/use-filters.ts
@@ -2,19 +2,13 @@ import { z } from "zod";
 
 import { AutoFormValueInfo, AutoFormValueType } from "@/components/auto-field";
 
-import {
-  CommentFilterKeys,
-  CommentFilters,
-  useCommentSelectAdapter,
-} from "@/entities/comment";
+import { CommentFilterKeys, CommentFilters } from "@/entities/comment";
 import { type FilterFeature } from "@/shared/libs";
 
 export function useCommentFilters(): FilterFeature<
   CommentFilterKeys,
   CommentFilters
 > {
-  const { getById, getItems, renderItem } = useCommentSelectAdapter();
-
   function buildFilterFields(): Record<CommentFilterKeys, AutoFormValueInfo> {
     return {
       idList: {
