+++ src/screens/transitionValidators/form-fields.tsx
@@ -1,5 +1,6 @@
-import { CheckboxFormField, TextFormField } from "@/components/form-fields";
 import { Control, FieldValues, Path } from "react-hook-form";
+
+import { CheckboxFormField, TextFormField } from "@/components/form-fields";
 
 export function TransitionValidatorFormFields<T extends FieldValues>({
   control,
