+++ src/screens/twinclassTriggers/form-fields.tsx
@@ -1,8 +1,10 @@
-import { CheckboxFormField, TextFormField } from "@/components/form-fields";
-import { FeaturerTypes } from "@/entities/featurer";
-import { FeaturerFormField } from "@/widgets/form-fields";
 import { Control } from "react-hook-form";
 import { z } from "zod";
+
+import { CheckboxFormField, TextFormField } from "@/components/form-fields";
+
+import { FeaturerTypes } from "@/entities/featurer";
+import { FeaturerFormField } from "@/widgets/form-fields";
 
 type TriggersFormValues = z.infer<any>;
