+++ src/widgets/crud-data-table/header/header.tsx
@@ -1,5 +1,5 @@
 import { Plus, RefreshCw, Search } from "lucide-react";
-import React, { ForwardedRef, useCallback, useEffect, useReducer } from "react";
+import React, { ForwardedRef, useCallback, useEffect } from "react";
 
 import { AutoFormValueInfo } from "@/components/auto-field";
