+++ src/widgets/form-fields/featurer/components/featurer-param/featurer-param-form.tsx
@@ -1,6 +1,9 @@
-import { FormFieldProps } from "@/components/form-fields/types";
-import { FormField } from "@/shared/ui";
 import { FieldValues } from "react-hook-form";
+
+import { FormFieldProps } from "@/components/form-fields";
+
+import { FormField } from "@/shared/ui";
+
 import {
   FeaturerParamFormItem,
   FeaturerParamFormItemProps,
